OSIS-6245 : Move files
